requirements.txt
ops>=2.14.1
pydantic==2.8.2
opentelemetry-exporter-otlp-proto-http==1.21.0
